requirements.txt
python-dotenv>=1.0.0
pyarrow>=14.0.0
duckdb>=0.9.0
apscheduler>=3.10.0 
loguru>=0.7.0
aiohttp>=3.8.0
aiofiles>=23.0.0
lxml>=4.9.0
python-dateutil>=2.8.0
tqdm>=4.64.0 
cachetools~=5.5.0
pandas~=2.2.3
numpy~=2.0.2
joblib~=1.4.2
pytz~=2024.2
colorama~=0.4.6